+++ NEWS.md
@@ -1,3 +1,10 @@
+# Version 1.4.4 (September 1st 2023)
+
+* Updated NCI Thesaurus - 23.08d
+* Updated CIViC data
+* Updated Mitelman database (20230803)
+* `get_biomarkers()` now exported as a main function
+
 # Version 1.4.3 (August 7th 2023)
 
 * NCI Thesaurus 23.07e